-menu icons layout logic is altered, now they won't slip away on big screens -money counter, money per sec and investment price does not show decimals anymore

+++ app/src/main/java/hu/uniobuda/nik/betherichest/InvestmentAdapter.java
@@ -88,7 +88,7 @@ public class InvestmentAdapter extends BaseAdapter {
             nameTextView.setTextColor(Color.parseColor("#760c07"));
         }
 
-        priceTextView.setText(String.valueOf(investment.getPrice()));
+        priceTextView.setText(String.valueOf((int)investment.getPrice()));
         dpsPerRankTextView.setText("DPS: " + String.valueOf(investment.getMoneyPerSecPerRank()));
         rankTextView.setText("Rank: " + String.valueOf(investment.getRank()));
         descriptionTextView.setText(investment.getDescription());
